+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - ...
-...
+### Added
+- add `use_embedded` kwarg to `TiffSlide.get_thumbnail`
 
 ## [0.1.1] - 2021-08-25
 ### Fixed